Moving layouts to a separate folder
